motiv8-be/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6
insightface>=0.7.3
onnxruntime>=1.16.0
torch>=2.0.0
opencv-python>=4.8.0
numpy>=1.24.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0
Pillow>=10.0.0
timm>=1.0.0
einops>=0.8.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
httpx>=0.24.0
boto3>=1.28.0
python-dotenv>=1.0.0
